Rename Admin.js component to AdminHome.js, update filename and and name in App.js
